requirements.txt
python-pptx
Pillow
numpy
scikit-learn
nltk
torch
torchvision